split into regular and legacy models. add some sharding support.
