+++ lib/stage-iteration.js
@@ -1,4 +1,5 @@
 var each = require('lodash.foreach');
+// var debug = require('debug')('mongodb-explain-plan-model:stage-iteration');
 
 var stageIterationMixin = {
   /**
@@ -44,8 +45,8 @@ var stageIterationMixin = {
   },
   /**
    * returns child stage or stages of current stage as array. If there are
-   * no more child stages, returns empty array []. Not supported for legacy
-   * mode.
+   * no more child stages, returns empty array []. Also works for sharded
+   * explain plans (where shards are children). Not supported for legacy mode.
    *
    * @param  {Object} stage   - stage to get children of.
    * @return {Array}          - array of child stages.
@@ -56,7 +57,13 @@ var stageIterationMixin = {
       return null;
     }
     stage = stage || this.rawExplainObject.executionStats.executionStages;
-    return stage.inputStage ? [stage.inputStage] : stage.inputStages;
+    if (stage.inputStage) {
+      return [stage.inputStage];
+    }
+    if (stage.executionStages) {
+      return [stage.executionStages];
+    }
+    return stage.shards || stage.inputStages || [];
   },
   /**
    * iterates over all stages and returns a depth-first pre-ordered array
@@ -68,7 +75,7 @@ var stageIterationMixin = {
   _getStageArray: function(root) {
     var result = [];
     var it = this._getStageIterator(root);
-    for (var stage = it.next(); stage !== null; stage = it.next()) {
+    for (var stage = it.next(); stage; stage = it.next()) {
       result.push(stage);
     }
     return result;
