+++ package.json
@@ -16,8 +16,12 @@
   "dependencies": {
     "ampersand-model": "^6.0.3",
     "lodash.assign": "^4.0.8",
+    "lodash.filter": "^4.4.0",
     "lodash.foreach": "^4.2.0",
+    "lodash.map": "^4.4.0",
     "lodash.set": "^4.1.0",
+    "lodash.some": "^4.4.0",
+    "lodash.uniq": "^4.3.0",
     "lodash.zip": "^4.0.0"
   },
   "devDependencies": {
